Address CR and fix updating duration should update pending
